Переход на docker
