+++ UI/Dockerfile
@@ -1,0 +1,10 @@
+# Указывает Docker использовать официальный образ python 3 с dockerhub в качестве базового образа
+FROM python:3.11.6-slim
+# Устанавливает переменную окружения, которая гарантирует, что вывод из python будет отправлен прямо в терминал без предварительной буферизации
+ENV PYTHONUNBUFFERED 1
+# Устанавливает рабочий каталог контейнера — "app"
+WORKDIR /app
+# Копирует все файлы из нашего локального проекта в контейнер
+ADD ./app/ /app
+# Запускает команду pip install для всех библиотек, перечисленных в requirements.txt
+RUN pip install -r requirements.txt
