--- UI/Pipfile
@@ -1,16 +1,0 @@
-[[source]]
-url = "https://pypi.python.org/simple"
-verify_ssl = true
-name = "pypi"
-
-[packages]
-django = "*"
-django-crispy-forms = "*"
-geopy = "*"
-geoip2 = "*"
-folium = "*"
-
-[dev-packages]
-
-[requires]
-python_version = "3.7"
