+++ UI/README.md
@@ -1,22 +1,31 @@
-## Обработка запросов и пользовательский интерфейс<br><br>
+<# GEO-Location-AI
 
-### Список файлов:<br>
-[Обработка и визуализация](processing.ipynb)<br>
+Система рекомендует пользователю рестораны и кафе по запросу.
+Рекомедации основаны на анализе отзывом посетителей, которые заранее предобработаны и 
+загружены в векторную БД.
 
-### Ход работы:<br>
+В системе реализована методика RAG:
 
-В системе реализована методика Retrieval-Augmented Generation (RAG), которая объединяет подходы поиска и генерации текста для получения максимально релевантных и полезных результатов. Процесс включает следующие этапы:
-1. **Семантический поиск релевантных отзывов**<br>
-    - На первом шаге используется биэнкодер, который преобразует текстовые данные (запрос пользователя и отзывы) в эмбеддинги. Это позволяет найти 10 наиболее подходящих отзывов, соответствующих запросу пользователя, на основе их семантического сходства.
+1. При помощи биэнкодера выполняется семантический поиск наиболее релевантных к запросу отзывов.
+2. Реранкенкер уточняет результаты и пересортирует их.
+3. 3 самых подходящих по отзывам точки выводятся на карту follium.
+4. Результат отбора вместе с промтом посылается в ChatGpt через OpenAI API.
+ChatGpt возвращает интересные факты и описание заведений. 
 
-2. **Уточнение результатов с помощью реранкинга**<br>
-    - Полученные отзывы передаются в реранкер, основанный на модели CrossEncoder. Реранкер анализирует пару "запрос–отзыв" и пересортировывает результаты с учетом их релевантности. Это позволяет повысить точность отбора.
+Инструкция по установке:
 
-3. **Визуализация лучших результатов на карте**<br>
-    - Три заведения, получившие наивысшие оценки после реранкинга, добавляются на интерактивную карту с помощью библиотеки Folium. Для каждого заведения указываются его местоположение, название, рейтинг и краткое описание.
+1. Скопировать каталог UI на сервер с docker compose.
+2. В файле app/geo.yml заполнить данные:
+- PROXY_API_KEY: ключ для API openai
+- PROXY_API_URL: адрес API openai
+- DJANGO_KEY: Django SECRET_KEY
+3. В папку db скопировать базу данных chromadb c отзывами
+4. В папку models загрузить модели:
+- deepvk/USER-bge-m3
+- DiTy/cross-encoder-russian-msmarco
+- spacy/ru_core_news_sm
+5. В папку geoip скачать карты mmdb. Скачать можно с сайта
+https://github.com/P3TERX/GeoLite.mmdb
+6. Запустить сервис командой docker-compose up
+7. Сервис будет доступен по адресу http://<адрес сервера>:9001
 
-4. **Генерация обзора и фактов через ChatGPT**<br>
-    - Сформированный список отзывов передается в модель ChatGPT через OpenAI API вместе с промтом, который задает инструкцию для генерации:
-        - Интересного факта или шутки о запросе.
-        - Краткого обзора заведения на основе отзывов.
-        - Итоговый текст возвращается пользователю, предоставляя информативное и увлекательное описание подходящих заведений.
