+++ UI/app/geo.yml
@@ -1,0 +1,3 @@
+PROXY_API_KEY: {fill}
+PROXY_API_URL: {fill}
+DJANGO_KEY: {fill}
