+++ UI/app/measurements/admin.py
@@ -1,0 +1,7 @@
+from django.contrib import admin
+from .models import Measurement
+
+# Register your models here.
+
+
+admin.site.register(Measurement)
