+++ UI/app/measurements/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class MeasurementsConfig(AppConfig):
+    name = 'measurements'
+    verbose_name = 'Distance between two locations'
