+++ UI/app/measurements/forms.py
@@ -1,0 +1,9 @@
+from django import forms
+from .models import Measurement
+
+
+class MeasurementModelForm(forms.ModelForm):
+    class Meta:
+        model = Measurement
+        fields = ('destination',)
+        labels = {'destination': 'Запрос'}
