+++ UI/app/measurements/templates/measurements/main.html
@@ -1,0 +1,55 @@
+{% extends 'base.html' %}
+{% load crispy_forms_tags %}
+
+{% block title %}
+        calclulate distance.
+{% endblock title %}
+
+
+
+{% block content %}
+
+
+<!-- Modal -->
+<!-- <div class="modal fade" id="exampleModal" tabindex="-1" aria-labelledby="exampleModalLabel" aria-hidden="true">
+  <div class="modal-dialog">
+    <div class="modal-content">
+      <div class="modal-header">
+        <h5 class="modal-title" id="exampleModalLabel">Distance to {{destination}} is: {{distance}}KM</h5>
+        <button type="button" class="close" data-dismiss="modal" aria-label="Close">
+          <span aria-hidden="true">&times;</span>
+        </button>
+      </div>
+      <div class="modal-body">
+        ...
+      </div>
+      <div class="modal-footer">
+        <button type="button" class="btn btn-secondary" data-dismiss="modal">Close</button>
+        <button type="button" class="btn btn-primary">Save changes</button>
+      </div>
+    </div>
+  </div>
+</div> -->
+
+
+<hr>
+
+<div class="input-group mb-3">
+<div style="width: 40%">
+<form action='' method='POST', autocomplete='off'>
+    {% csrf_token %}
+    {{form|crispy}}
+    <button type='submit' class='btn btn-primary'>Искать</button>
+    </form>
+</div>
+<textarea rows="3" style="margin-left: 5%; width: 55%;" disabled>
+{{text}}
+</textarea>
+</div>
+    <br>
+
+
+
+{{map|safe}}
+
+{% endblock content %}
