+++ UI/app/measurements/utils.py
@@ -1,0 +1,38 @@
+#helper functions
+from django.contrib.gis.geoip2 import GeoIP2
+
+def get_geo(ip):
+    g = GeoIP2()
+    country = g.country(ip)
+    city = g.city(ip)
+    lat, lon = g.lat_lon(ip)
+    return country, city, lat, lon
+
+def get_center_coordinates(latA, longA, latB=None, longB=None):
+    cord = (latA, longA)
+
+    if latB:
+        cord = [(latA+latB)/2, (longA+longB)/2]
+
+    return cord
+
+
+def get_zoom(distance):
+    """ Determine the zoom in value based on the distance. """
+    if distance <= 100:
+        return 8
+    elif distance > 100 and distance <=7000:
+        return 4
+    else:
+        return 1
+        
+
+def get_client_ip(request):
+    """Function to get the users ip address"""
+    
+    x_forwarded_for = request.META.get('HTTP_X_FORWARDED_FOR')
+    if x_forwarded_for:
+        ip = x_forwarded_for.split(',')[0]
+    else:
+        ip = request.META.get('REMOTE_ADDR')
+    return ip
