--- UI/distance_proj/asgi.py
@@ -1,16 +1,0 @@
-"""
-ASGI config for distance_proj project.
-
-It exposes the ASGI callable as a module-level variable named ``application``.
-
-For more information on this file, see
-https://docs.djangoproject.com/en/3.0/howto/deployment/asgi/
-"""
-
-import os
-
-from django.core.asgi import get_asgi_application
-
-os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'distance_proj.settings')
-
-application = get_asgi_application()
