--- UI/distance_proj/urls.py
@@ -1,8 +1,0 @@
-
-from django.contrib import admin
-from django.urls import path, include
-
-urlpatterns = [
-    path('admin/', admin.site.urls),
-    path('', include('measurements.urls', namespace='measurements'))
-]
