+++ UI/geoip/read
@@ -1,0 +1,2 @@
+Cкачать сюда карты mmdb. Скачать можно с сайта
+https://github.com/P3TERX/GeoLite.mmdb
