--- UI/measurements/admin.py
@@ -1,7 +1,0 @@
-from django.contrib import admin
-from .models import Measurement
-
-# Register your models here.
-
-
-admin.site.register(Measurement)
