--- UI/measurements/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class MeasurementsConfig(AppConfig):
-    name = 'measurements'
-    verbose_name = 'Distance between two locations'
