--- UI/requirements.txt
@@ -1,24 +1,0 @@
-asgiref==3.2.10
-branca==0.4.1
-certifi==2020.6.20
-chardet==3.0.4
-Django==3.0.7
-django-crispy-forms==1.9.1
-folium
-geographiclib==1.50
-geoip2==3.0.0
-geopy==1.22.0
-idna==2.9
-Jinja2==2.11.2
-MarkupSafe==1.1.1
-maxminddb==1.5.4
-numpy
-pytz==2020.1
-requests==2.24.0
-sqlparse==0.3.1
-urllib3==1.25.9
-langchain-community
-sentence-transformers
-openai
-spacy
-pyyaml
